add tabs (first pass)
