+++ src/aics-image-viewer/components/ControlPanel/styles.css
@@ -27,11 +27,11 @@
     }
 
     .btn-collapse {
-      > .anticon {
+      svg {
         transform: rotate(0deg);
         transition: transform 0.2s linear;
       }
-      &.btn-collapse-collapsed > .anticon {
+      &.btn-collapse-collapsed svg {
         transform: rotate(180deg);
       }
     }
